requirements.txt
exifread==3.3.0
geopy==2.4.1
Pillow==7.0.0
Pillow==11.2.1
pillow_heif==0.18.0
